Gi tests (#237)
* CMake: Reworked install script.
* CMake: Updated projects versions.

* CastorUtils: Reworked Point data to be able to use x, y, and z.
* CastorUtils: Added RangeSequence.
* CastorUtils: Removed inappropriate operators in Pixel.
* CastorUtils: Added Perlin and Simplex noise algorithms.
* CastorUtils: Reworked ArrayView and FlagCombination.
* CastorUtils: Added == and != operators to BoundingBox and BoundingSphere

* Castor3D: Crash fixes on mesh loading failure.
* Castor3D: Added create functions for TextureLayout.
* Castor3D: Reworked post effects creation.
* Castor3D: Reworked GaussianBlur, to be able to use it externally
* Castor3D: Created customisable Combine pass.
* Castor3D: Moved independant render passes to Render/Passes folder.
* Castor3D: Fixed double image sampling.
* Castor3D: Fixed ignored light intensity.
* Castor3D: Reintroduced depth pre-pass.
* Castor3D: Uniformisation of geometry buffers' data.
* Castor3D: SSAO is now fully configurable through UBO (and through GuiCommon).
* Castor3D: Removed setDebugObjectName
* Castor3D: CombinePass no longer owns the shaders
* Castor3D: removed useless lightType parameter.
* Castor3D: Added spitScale member to directional lights, to fix RSM with cascades.
* Castor3D: Implemented LightInjection
* Castor3D: Finished basic implementation of Light Propagation Volumes
* Castor3D: Renamed LightInjectionUbo to LpvConfigUbo
* Castor3D: Fixed light propagation pass.
* Castor3D: Implemented cascaded LPV
* Castor3D: Fixed use of normal map in SAO.
* Castor3D: SAO min radius is now configurable via the scene file.
* Castor3D: Reworked a bit pipeline visitors
* Castor3D: Reworked GaussianBlur so it uses the source image dimensions
* Castor3D: Reworked GpInfoUbo so only one instance exists
* Castor3D: Moved SSGI and SSAO passes out of Technique.
* Castor3D: Added intermediate GBuffer class.
* Castor3D: Reworked TextureLayout, it now has the placeholders for on demand image views, for each layer/slice and for each mip level.
* Castor3D: Uniformisation of TextureLayout::getDeafultView
* Castor3D: Cleaned up a bit light shaders.
* Castor3D: Reworked RenderQuad and CombinePass.
* Castor3D: Moved TextureFlag to MaterialModule.hpp
* Castor3D: Updated descriptor filling to use a restricting mask.
* Castor3D: Reworked the way TextureFlags are handled
* Castor3D: RenderQuad is now a castor::Named.
* Castor3D: Fixed an issue in MipView.
* Castor3D: Uniformisation for shadow maps GBuffer creation/use.
* Castor3D: Extended lighting model to use light buffers in shadow mapping.
* Castor3D: Fixed combination shader in RenderTarget.
* Castor3D: Fixed color blend state creation in shadow map passes.
* Castor3D: Moved back RenderPass functions to ShadowMapPass and children.
* Castor3D: ShadowMap depth is now stored in w component of its map.
* Castor3D: Reworked flux computation.
* Castor3D: Reworked TextureView, to hold both sampled and target views
* Castor3D: Naive implementation of RSM is done.
* Castor3D: Fixed shadow generation for point and spot lights.
* Castor3D: Move RSM config UBO to its own file.
* Castor3D: Few more fixes.
* Castor3D: Implemented interpolation step to optimize a bit RSM.
* Castor3D: First try at voxelizing the scene.
* Castor3D: First try at voxelizing the scene.
* Castor3D: Removed unused camera parameter to shadow map passes

* GuiCommon: Reworked ShaderDialog
* GuiCommon: Castor3D version is now displayed on splash screen.
* GuiCommon: Changed debug option names.
* GuiCommon: Added a way to view intermediate images.

* CastorViewer: Added a perspective with scene panes displayed

+++ include/Core/Castor3D/Buffer/UniformBuffer.hpp
@@ -299,7 +299,7 @@ namespace castor3d
 			, RenderPassTimer const & timer
 			, uint32_t index )const
 		{
-			data.resize( m_elemCount * m_elemSize );
+			data.resize( size_t( m_elemCount ) * m_elemSize );
 			return download( stagingBuffer
 				, queue
 				, commandPool
