+++ test/CastorUtils/CastorUtilsPixelFormatTest.cpp
@@ -307,7 +307,7 @@ namespace
 		{
 			Size size = { 16, 16 };
 			std::vector< uint8_t > buffer;
-			size_t count = size.getWidth() * size.getHeight() * PixelDefinitions< PFSrc >::Size;
+			size_t count = size_t( size.getWidth() ) * size.getHeight() * PixelDefinitions< PFSrc >::Size;
 			buffer.resize( count );
 			uint8_t value = 0;
 			Pixel< PFSrc > pixel;
@@ -347,7 +347,7 @@ namespace
 		{
 			Size size = { 16, 16 };
 			std::vector< uint8_t > buffer;
-			size_t count = size.getWidth() * size.getHeight() * PixelDefinitions< PFSrc >::Size;
+			size_t count = size_t( size.getWidth() ) * size.getHeight() * PixelDefinitions< PFSrc >::Size;
 			buffer.resize( count );
 			uint32_t depth = 0;
 			uint8_t stencil = 0;
@@ -381,7 +381,7 @@ namespace
 		{
 			Size size = { 16, 16 };
 			std::vector< uint8_t > buffer;
-			size_t count = size.getWidth() * size.getHeight() * PixelDefinitions< PFSrc >::Size;
+			size_t count = size_t( size.getWidth() ) * size.getHeight() * PixelDefinitions< PFSrc >::Size;
 			buffer.resize( count );
 			uint32_t depth = 0;
 			uint8_t stencil = 0;
@@ -416,7 +416,7 @@ namespace
 		{
 			Size size = { 16, 16 };
 			std::vector< uint8_t > buffer;
-			size_t count = size.getWidth() * size.getHeight() * PixelDefinitions< PFSrc >::Size;
+			size_t count = size_t( size.getWidth()) * size.getHeight() * PixelDefinitions< PFSrc >::Size;
 			buffer.resize( count );
 			uint8_t value = 0;
 			Pixel< PFSrc > pixel;
